GENAI/Langchain/others/Chains/code/requirements.txt
langchain-openai
langchain_anthropic
langchain-community
langchain-core

python-dotenv